Removing lombok
It was perhaps a little premature on my part to add lombok and start switching things over to it. While I do believe it's a worthy addition, it wasn't for me to add it without discussing it first. Luckily, most of my additions are APIs and not data classes, so not much needed to be reverted.