Fixed potential processing leak
ClickableButton's `onRemovedFromInventory()` method was added to help make buttons more eventful, but also to fix an underlying issue with AnimatedClickable, whereby it knew when it start animating but not when to stop, so there was a possibility of the server being significantly slowed down by animated clickables that have long since disappeared, which may mean continuously thrown errors or perhaps keeping the ClickableInventory from being garbage collected.

However, I thought that this could be called when its ClickableInventory is closing, but that would mean the button has no way to distinguish between being purposefully removed and routine cleanup. And so `onInventoryClose()` has been added to fix that issue.