Iteration :: hidden bug fixed
I've deprecated but also changed the signature of `Iteration.isNullOrEmpty(T... array)` to `Iteration.isNullOrEmpty(T[] array)` which may cause problems, but not fixing this would cause more problems down the line due to hidden bugs. Basically, for some things, such as Maps, you may forget (as I did) that they aren't collections, but will still look correct to the compiler because it's using the generic array version instead of the collection version, which bypasses the actual check you're doing about whether the Map is null or empty, which for that I've created a specific check in MapUtils. I've deprecated the generic array version because a better version exists thanks to `ArrayUtils`.